+++ GeometryDashAPI/Levels/GameObjects/IBlock.cs
@@ -9,7 +9,7 @@ public interface IBlock : IGameObject
     float PositionY { get; set; }
     bool HorizontalReflection { get; set; }
     bool VerticalReflection { get; set; }
-    int Rotation { get; set; }
+    float Rotation { get; set; }
     bool Glow { get; set; }
     int LinkControl { get; set; }
     short EditorL { get; set; }
